requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.20
opencv-python>=4.8.1.78
numpy>=1.24.3
Pillow>=10.0.1
streamlit>=1.28.0
scikit-learn>=1.3.2
matplotlib>=3.7.3
pandas>=2.1.1
tensorflow>=2.13.0